🚚 Rename to NavLinkBarForWelcome as a part of welcome page
